♻️ Separate data fetching into a custom hook
